%[ER: 231]% Merge with branch ER231_development.

+++ trunk/CEP/tinyCEP/src/Sel_RoundRobin.cc
@@ -48,7 +48,7 @@ Selector* Sel_RoundRobin::clone() const
 unsigned int Sel_RoundRobin::getNext()
 {
   itsCurrentSelection++;
-  if (itsCurrentSelection >= itsNOptions)
+  if (itsCurrentSelection >= (int)itsNOptions)
   {
     itsCurrentSelection = 0;
   }
